third_party/WebKit/LayoutTests/http/tests/htmlimports/import-onerror.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script>
window.jsTestIsAsync = true;
testRunner.waitUntilDone();

var givenLoadEventTargets = [];
function handleError(event)
{
    givenLoadEventTargets.push(event.target.id);
    if (givenLoadEventTargets.length == 3) {
        shouldBe("['invalid', 'notFound', 'notFoundShared']", "givenLoadEventTargets.sort()");
        finishJSTest();
    }
}
</script>
<link id="notFound" rel="import" href="notfound.html" onerror="handleError(event)">
<link id="notFoundShared" rel="import" href="notfound.html" onerror="handleError(event)">
<link id="invalid" rel="import" href="" onerror="handleError(event)">
</head>
<body></body>
</html>
